+++ tests/test_agate.py
@@ -13,12 +13,12 @@ import agate
 class TestCSV(unittest.TestCase):
     def test_agate(self):
         if six.PY2:
-            self.assertIs(agate.reader, agate.csv_py2.reader)
-            self.assertIs(agate.writer, agate.csv_py2.writer)
-            self.assertIs(agate.DictReader, agate.csv_py2.DictReader)
-            self.assertIs(agate.DictWriter, agate.csv_py2.DictWriter)
+            self.assertIs(agate.csv.reader, agate.csv_py2.reader)
+            self.assertIs(agate.csv.writer, agate.csv_py2.writer)
+            self.assertIs(agate.csv.DictReader, agate.csv_py2.DictReader)
+            self.assertIs(agate.csv.DictWriter, agate.csv_py2.DictWriter)
         else:
-            self.assertIs(agate.reader, agate.csv_py3.reader)
-            self.assertIs(agate.writer, agate.csv_py3.writer)
-            self.assertIs(agate.DictReader, agate.csv_py3.DictReader)
-            self.assertIs(agate.DictWriter, agate.csv_py3.DictWriter)
+            self.assertIs(agate.csv.reader, agate.csv_py3.reader)
+            self.assertIs(agate.csv.writer, agate.csv_py3.writer)
+            self.assertIs(agate.csv.DictReader, agate.csv_py3.DictReader)
+            self.assertIs(agate.csv.DictWriter, agate.csv_py3.DictWriter)